feat: 레포트 작업 중

+++ src/main/java/com/example/cooing/domain/report/dto/TotalInfoResponseDto.java
@@ -6,6 +6,6 @@ import lombok.Getter;
 @Getter
 @Builder
 public class TotalInfoResponseDto {
-  private final Long totalWordNum;
+  private final Integer totalWordNum;
   private final String mostUseWord;
 }
